+++ Dart Challange/unit_test_2/bin/unit_test_2.dart
@@ -1,0 +1,5 @@
+import 'package:unit_test_2/get_list_item_in_two_list.dart' as unit_test_2;
+
+void main(List<String> arguments) {
+  unit_test_2.getListItemsInTwoList(listNumber1: [], listNumber2: []);
+}
